requirements.txt
pillow==7.2.0
scikit-image==0.17.2
